data(worldbank_wdi): move variable meta creation from grapher to garden step
