+++ etl/steps/data/garden/worldbank_wdi/2022-05-26/wdi.variable_mapping.json
@@ -9,6 +9,8 @@
   },
   "series_changes_2021": {
     "change_in_description": {
+      "Contraceptive prevalence, any methods (% of women ages 15-49)": "Contraceptive prevalence, any method (% of married women ages 15-49)",
+      "Contraceptive prevalence, modern methods (% of women ages 15-49)": "Contraceptive prevalence, any modern method (% of married women ages 15-49)",
       "Net official development assistance and official aid received (constant 2015 US$)": "Net official development assistance and official aid received (constant 2018 US$)",
       "Net official aid received (constant 2015 US$)": "Net official aid received (constant 2018 US$)",
       "Net official development assistance received (constant 2015 US$)": "Net official development assistance received (constant 2018 US$)",
